+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 * Added relic: Spicy Sausage
 * Fix: Water orb shouldn't be affected by focus
+* Added relics: Treasure Map
 
 ### v1.4.1
 * In multi monster fights, the left monsters don't apply vulnerable.